binder: fix binder_ref delete-before-add race
During binder_ref creation, the proc lock is released
once the ref is added to the binder_proc's rb_tree and
the target node's proc lock is acquired to add the ref
to the node's ref list.

It is possible for another thread to use the binder_ref
and delete it before it is added to the node's ref list.
This resulted in a crash. Fixed by adding an implicit
weak ref when during binder_get_ref* and adding a
binder_put_ref() which must be used when a ref
obtained via binder_get_ref* is no longer needed.

Bug: 35764014
Change-Id: I4758cf5afcca61b02008c04afb95c3e0991bb912
Test: tested manually
Signed-off-by: Todd Kjos <tkjos@google.com>