misc: Fix some memory leaks
Squash of the following memory leak fixes:

net: ipc_router: Fix memory leaks when releasing a remote port

In ipc_router_set_conn(), memory is allocated to a local port for the
conn_info linked list, but the memory for the local ports is never
freed when the remote port is released.

Delete the local ports and free all of their memory when releasing a remote
port to fix the memory leaks. This is just like how memory is freed for the
remote_tx ports.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

soc: qcom: watchdog_v2: Fix memory leaks when memory_dump_v2 isn't built

When the memory_dump_v2.c driver isn't built (CONFIG_MSM_MEMORY_DUMP_V2=n),
an inline version of msm_dump_data_register() will be used that does
nothing. This means that all memory allocated with the intention of going
to msm_dump_data_register() will be leaked.

Fix this by ignoring the entire code block when CONFIG_MSM_MEMORY_DUMP_V2
is disabled.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

soc: qcom: glink: Fix memory leak in driver probe

Memory is allocated to arr32 so that it can temporarily store data parsed
from the device tree, but the memory isn't freed after it's done being
used.

Free the memory allocated to arr32 to fix the leak.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

power: bq27541: Fix memory leak in driver probe

Memory is allocated for a name that isn't used anywhere, so just remove it
entirely to fix the leak.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

msm: camera: isp: Fix memory leaks in vfe probe

Memory is needlessly allocated to vfe_parent_dev and vfe_parent_dev->common_sd
and never freed.

Since the allocated memory is not used anywhere outside of vfe_probe(),
and since it's unnecessary, just remove the allocations entirely and use a
stack variable instead.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

cpufreq: stats: Fix memory leaks when updating stats table

The address for an element in the per-cpu cpufreq_stats_table variable
is overwritten in cpufreq_stats_update_policy_cpu(), but the memory
allocated at the address that gets overwritten is not freed beforehand.

Free the allocated memory beforehand to fix the memory leaks.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

msm: camera: sensor: Fix memory leaks when sensor probe fails

The power settings for a camera sensor can be parsed and have its memory
never freed in the error path if part of the probe sequence fails after
the power settings are parsed.

Free the power settings in the error paths to fix the memory leaks.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

input: synaptics: fw_update: Fix various memory leaks

Memory allocated to fwu->read_config_buf is not freed in both the success
and error paths. Fix it.

Signed-off-by: Sultanxda <sultanxda@gmail.com>

Signed-off-by: Alex Naidis <alex.naidis@linux.com>
Signed-off-by: dd3boh <dade.garberi@gmail.com>

msm: camera: sensor: Fix memory leaks when sensor probe fails

The power settings for a camera sensor can be parsed and have its memory
never freed in the error path if part of the probe sequence fails after
the power settings are parsed.

Free the power settings in the error paths to fix the memory leaks.

@DD3Boh: Reapply on camera_v2z

Signed-off-by: Sultanxda <sultanxda@gmail.com>
Signed-off-by: dd3boh <dade.garberi@gmail.com>